bin: add interactive mode
I'm slightly concerned that this will hang in CI, the node readline
module is a bit touchy when stdin is not a tty.

If it's a huge problem, can just make the timeout signal make the test
pass rather than fail, even though that's really not ideal.  It seems
really reliable when it is actually interactive, though.