Only call directory removal method on actual dirs
This prevents all symbolic link following, and also cuts down on the
system calls quite considerably, while still preventing the "unlinked
directory" issue on SunOS, with only the cost of a single lstat at the
start of the process.

Also added test of mid-stream AbortSignal triggering in sync rimrafs,
because this is possible if called from a filter method.

Fix: #259